Revert "Move GCPU to the gcpu directory since names must match"
This reverts commit a51f50699d0d1ab602a71226687cf053a9ade27d.
