+++ src/ngraph/runtime/generic_cpu/gcpu_executable.hpp
@@ -26,9 +26,9 @@
 #include "ngraph/ops.hpp"
 #include "ngraph/runtime/aligned_buffer.hpp"
 #include "ngraph/runtime/backend.hpp"
-#include "ngraph/runtime/gcpu/kernel/broadcast.hpp"
-#include "ngraph/runtime/gcpu/kernel/dot.hpp"
-#include "ngraph/runtime/gcpu/kernel/reshape.hpp"
+#include "ngraph/runtime/generic_cpu/kernel/broadcast.hpp"
+#include "ngraph/runtime/generic_cpu/kernel/dot.hpp"
+#include "ngraph/runtime/generic_cpu/kernel/reshape.hpp"
 #include "ngraph/runtime/host_tensor.hpp"
 #include "ngraph/runtime/interpreter/int_executable.hpp"
 #include "ngraph/runtime/reference/abs.hpp"
