+++ src/ngraph/runtime/generic_cpu/kernel/dot.hpp
