+++ src/ngraph/runtime/generic_cpu/unit_test.manifest
